Code cleaned_home section modified
